move shortlinkinfo to /api
