+++ lib/handlers/api.js
@@ -26,13 +26,15 @@ const express = require('express'),
     _ = require('underscore'),
     AsmDocsApi = require('./asm-docs-api'),
     FormatterHandler = require('./formatting'),
-    utils = require('../utils');
+    utils = require('../utils'),
+    clientStateNormalizer = require('../clientstate-normalizer').ClientStateNormalizer;
 
 class ApiHandler {
-    constructor(compileHandler, ceProps) {
+    constructor(compileHandler, ceProps, storageHandler) {
         this.compilers = [];
         this.languages = [];
         this.usedLangIds = [];
+        this.storageHandler = storageHandler;
         this.handle = express.Router();
         this.handle.use((req, res, next) => {
             res.header("Access-Control-Allow-Origin", "*");
@@ -50,6 +52,34 @@ class ApiHandler {
         const formatter = new FormatterHandler(ceProps);
         this.handle.post('/format/:tool', formatter.formatHandler.bind(formatter));
         this.handle.get('/formats', formatter.listHandler.bind(formatter));
+
+        this.handle.get('/shortlinkinfo/:id', this.shortlinkInfoHandler.bind(this))
+    }
+
+    shortlinkInfoHandler(req, res, next) {
+        const id = req.params.id;
+        this.storageHandler.expandId(id)
+            .then(result => {
+                const config = JSON.parse(result.config);
+
+                if (config.content) {
+                    const normalizer = new clientStateNormalizer();
+                    normalizer.fromGoldenLayout(config);
+
+                    res.set('Content-Type', 'application/json');
+                    res.end(JSON.stringify(normalizer.normalized));
+                } else {
+                    res.set('Content-Type', 'application/json');
+                    res.end(JSON.stringify(config));
+                }
+            })
+            .catch(err => {
+                logger.warn(`Exception thrown when expanding ${id}: `, err);
+                next({
+                    statusCode: 404,
+                    message: `ID "${id}" could not be found`
+                });
+            });
     }
 
     handleLanguages(req, res) {
